Fix return types for query 3

+++ src/main/java/com/ldbc/driver/workloads/ldbc/snb/bi/LdbcSnbBiQuery3TagEvolution.java
@@ -21,12 +21,12 @@ public class LdbcSnbBiQuery3TagEvolution extends Operation<List<LdbcSnbBiQuery3T
         this.limit = limit;
     }
 
-    public long year()
+    public int year()
     {
         return year;
     }
 
-    public long month()
+    public int month()
     {
         return month;
     }
